+++ src/app/administration/team-members/team-members-container/team-members-container.component.html
@@ -3,14 +3,17 @@
     mat-button
     class="create-new-button"
     color="primary"
-    [disabled]="!isUserTeamAdmin$ && !(userIsLZAdmin$ | async)"
+    [disabled]="!userIsTeamAdmin && !(userIsLZAdmin$ | async)"
     (click)="addNewTeamMember()"
   >
     + Add New Team Member
   </button>
   <app-grid-view-switch class="view-settings-bar" [gridViewName]="gridViewOptionsName"></app-grid-view-switch>
 </div>
-<app-team-members-grid *ngIf="!(isGridViewEnabled$ | async); else listview"></app-team-members-grid>
+<app-team-members-grid
+  *ngIf="isGridViewEnabled$ | async; else listview"
+  [userIsTeamAdmin]="userIsTeamAdmin"
+></app-team-members-grid>
 <ng-template #listview>
-  <app-team-members-list></app-team-members-list>
+  <app-team-members-list [userIsTeamAdmin]="userIsTeamAdmin"></app-team-members-list>
 </ng-template>
