+++ src/app/administration/team-members/team-members-container/team-members-container.component.ts
@@ -8,7 +8,8 @@ import { selectGridViewSwitchOptions } from '@app/shared/grid-view-switch/grid-v
 import { select, Store } from '@ngrx/store';
 import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
-import { selectUserIsLZAdmin, selectUserIsTeamAdmin } from '@app/login/login.reducer';
+import { selectUserId, selectUserIsLZAdmin } from '@app/login/login.reducer';
+import { TeamMember } from '../team-members.model';
 
 @Component({
   selector: 'app-team-members-container',
@@ -16,16 +17,23 @@ import { selectUserIsLZAdmin, selectUserIsTeamAdmin } from '@app/login/login.red
   styleUrls: ['./team-members-container.component.scss'],
 })
 export class TeamMembersContainerComponent implements OnInit {
-  gridViewOptionsName: GridViewSwitchViewsNames = GridViewSwitchViewsNames.teamMembers;
   currentGridViewOption$: Observable<string>;
-  isUserTeamAdmin$: Observable<boolean>;
+  gridViewOptionsName: GridViewSwitchViewsNames = GridViewSwitchViewsNames.teamMembers;
+  loggedUserId: number;
+  TeamMembers: TeamMember[];
   userIsLZAdmin$: Observable<boolean>;
+  userIsTeamAdmin: boolean;
+
   constructor(private store: Store<any>, private route: ActivatedRoute, private router: Router) {}
 
   ngOnInit() {
     this.currentGridViewOption$ = this.store.pipe(select(selectGridViewSwitchOptions(this.gridViewOptionsName)));
     this.userIsLZAdmin$ = this.store.pipe(select(selectUserIsLZAdmin));
-    this.isUserTeamAdmin$ = this.store.pipe(select(selectUserIsTeamAdmin));
+    this.store.pipe(select(selectUserId)).subscribe((userId) => (this.loggedUserId = userId));
+    this.TeamMembers = this.route.snapshot.data['teamMembers'] as TeamMember[];
+    this.userIsTeamAdmin = this.TeamMembers.some(
+      (teamMember) => teamMember.userId === this.loggedUserId && teamMember.isTeamAdmin
+    );
   }
 
   get isGridViewEnabled$(): Observable<boolean> {
