+++ src/app/administration/team-members/team-members-grid/team-members-grid-card/team-members-grid-card.component.scss
