+++ src/app/administration/team-members/team-members-grid/team-members-grid.component.html
@@ -1,53 +1,14 @@
-<div class="mat-elevation-z8">
-  <table mat-table [dataSource]="dataSource" matSort>
-    <!-- ID Column -->
-    <ng-container matColumnDef="id">
-      <th mat-header-cell *matHeaderCellDef mat-sort-header>ID</th>
-      <td mat-cell *matCellDef="let row">{{ row.id }}</td>
-    </ng-container>
-
-    <!-- Application Column -->
-    <ng-container matColumnDef="isActive">
-      <th mat-header-cell *matHeaderCellDef mat-sort-header>Active</th>
-      <td mat-cell *matCellDef="let row">
-        <mat-chip [color]="statusColor(row.isActive)" selected>{{ row.isActive ? 'Active' : 'Inactive' }}</mat-chip>
-      </td>
-    </ng-container>
-
-    <!-- Environment Column -->
-    <ng-container matColumnDef="isTeamAdmin">
-      <th mat-header-cell *matHeaderCellDef mat-sort-header>Team Admin</th>
-      <td mat-cell *matCellDef="let row">
-        <mat-chip *ngIf="row.isTeamAdmin; else notTeamAdmin" color="primary" selected>
-          {{ 'True' }}
-        </mat-chip>
-        <ng-template #notTeamAdmin>
-          <mat-chip>False</mat-chip>
-        </ng-template>
-      </td>
-    </ng-container>
-
-    <!-- Description Column -->
-    <ng-container matColumnDef="userInfo">
-      <th mat-header-cell *matHeaderCellDef mat-sort-header>User</th>
-      <td mat-cell *matCellDef="let row">{{ row.userInfo }}</td>
-    </ng-container>
-
-    <!-- Actions Column -->
-    <ng-container matColumnDef="actions">
-      <th mat-header-cell *matHeaderCellDef>Actions</th>
-      <td mat-cell class="nowrap" *matCellDef="let row" [style.color]="row.color">
-        <a mat-icon-button matTooltip="View user details" (click)="viewUser(row.id)">
-          <mat-icon aria-hidden="false" aria-label="View user icon">remove_red_eye</mat-icon>
-        </a>
-        <a *ngIf="isTeamAdmin()" mat-icon-button matTooltip="Edit user details" (click)="editUser(row.id)">
-          <mat-icon aria-hidden="false" aria-label="Edit user icon">edit</mat-icon>
-        </a>
-      </td>
-    </ng-container>
-    <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
-    <tr mat-row *matRowDef="let row; columns: displayedColumns"></tr>
-  </table>
-
-  <mat-paginator [pageSizeOptions]="[5, 10, 25, 100]"></mat-paginator>
+<div class="grid-container solution-grid-container">
+  <mat-grid-list
+    [cols]="(layout$ | async).gridCols"
+    [rowHeight]="(layout$ | async).solutionGridItemHeight"
+    gutterSize="24px"
+  >
+    <mat-grid-tile *ngFor="let teamMember of teamMembers" colspan="1" rowspan="1">
+      <app-team-members-grid-card
+        [userIsTeamAdmin]="userIsTeamAdmin"
+        [teamMember]="teamMember"
+      ></app-team-members-grid-card>
+    </mat-grid-tile>
+  </mat-grid-list>
 </div>
