+++ src/app/administration/team-members/team-members-grid/team-members-grid.component.scss
@@ -1,32 +1,20 @@
-.mat-elevation-z8 {
-  margin: 1rem;
-  margin-top: 0;
-  overflow: auto;
+.grid-container {
+  margin: 0 1rem;
 }
 
-.pagination-container {
-  display: flex;
-  justify-content: flex-end;
-}
-
-table {
+.create-new-button {
+  height: 100%;
   width: 100%;
-}
+  border-radius: 10px;
+  border: 1px dashed var(--dark-grey-2);
+  font-size: 1.5rem;
+  padding: 0 2rem;
 
-.mat-form-field {
-  font-size: 14px;
-  width: 100%;
-}
-
-.actions-top-bar {
-  display: flex;
-  margin-bottom: 0.5rem;
-}
-
-.action-button {
-  min-width: fit-content;
-}
-
-.mat-chip {
-  border-radius: 4px;
+  .mat-icon {
+    line-height: 3rem;
+    font-size: 2rem;
+    height: 3rem;
+    width: 3rem;
+    margin-left: -2rem;
+  }
 }
