+++ src/app/administration/users/users-home/users-home-grid/users-home-grid-card/users-home-grid-card.component.html
@@ -48,6 +48,6 @@
   </mat-card-content>
   <mat-card-actions *ngIf="active">
     <a mat-button routerLink="view" [queryParams]="{ id: user.id }"> View </a>
-    <a mat-button routerLink="edit" [queryParams]="{ id: user.id }"> Editasd </a>
+    <a mat-button routerLink="edit" [queryParams]="{ id: user.id }"> Edit </a>
   </mat-card-actions>
 </mat-card>
