+++ src/app/mission-control/solutions/solutions-view/solutions-view.component.html
@@ -65,11 +65,7 @@
   </mat-tab>
   <mat-tab label="Team members">
     <div class="tab-content">
-      <app-team-members-grid
-        *ngIf="teamMembers$ !== undefined"
-        [_teamMembers]="teamMembers$"
-        [isAddTeamMemberButtonVisible]="false"
-      ></app-team-members-grid>
+      <app-team-members-list *ngIf="teamMembers$ !== undefined" [TeamMembers$]="teamMembers$"></app-team-members-list>
     </div>
   </mat-tab>
 </mat-tab-group>
